pyproject.toml
[project]
name = "sandbox"
version = "0.1.0"
description = "Structured learning and experimentation environment for data systems."
requires-python = ">=3.14"
dependencies = [
  "polars",
  "pyarrow",
  "pandas",
  "duckdb",
  "deltalake",
  "prefect",
  "pulumi",
  "rich",
  "tabulate",
  "structlog",
  "pyyaml",
  "pydantic",
]

[dependency-groups]
dev = [
  "pre-commit",
  "black",
  "ruff",
  "pytest",
  "ipykernel",
]

[tool.black]
line-length = 100
target-version = ["py312"]
skip-string-normalization = true

[tool.ruff]
line-length = 100
target-version = "py312"
lint.select = ["E", "F", "I", "UP", "B", "PL", "W"]
lint.ignore = ["E203", "PLR2004"] # PLR2004 is ignoring hardcoded literals in f-strings for now
fix = true

[tool.pytest.ini_options]
addopts = "-q"
python_files = ["test_*.py"]
